# api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8082","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/shorten":{"post":{"tags":["url-controller"],"operationId":"shortenUrl","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UrlRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UrlResponse"}}}}}}},"/registration":{"post":{"tags":["registration-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/login":{"post":{"tags":["login-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginCredentials"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/{code}":{"get":{"tags":["url-controller"],"operationId":"redirectToFullUrl","parameters":[{"name":"code","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/private/my-links":{"get":{"tags":["url-controller"],"operationId":"getUserLinks","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ShortUrlResponse"}}}}}}}}},"components":{"schemas":{"UrlRequest":{"type":"object","properties":{"url":{"type":"string"},"customCode":{"type":"string"}}},"UrlResponse":{"type":"object","properties":{"shortUrl":{"type":"string"}}},"RegistrationRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"}}},"LoginCredentials":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"ShortUrlResponse":{"type":"object","properties":{"code":{"type":"string"},"longUrl":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"clickCount":{"type":"integer","format":"int64"}}}},"securitySchemes":{"bearerAuth":{"type":"http","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}